+++ R/gpt_annot_melt.R
@@ -1,5 +1,5 @@
 gpt_annot_melt <- function(res_coded){
-  id.vars <- grep("justification|phenotype|hpo_id|pheno_count",
+  id.vars <- grep("justification|phenotype|hpo_name|hpo_id|pheno_count",
                   names(res_coded$annot),
                   value = TRUE)
   dat <- data.table::melt.data.table(